make HashValue() support integers only
MySQL allows constructs like "WHERE ( transaction_id = '985999381' )"
even if transaction_id is an interger column. To process
such values properly in HashValue function we either need
to pass it a column type or derive it from somewhere else.

Instead we remove support for sharding on string types
for now. So sharding type "shard" means that unit is
expected as a sharding key in a form of int, string, or
[]byte.

We can add support for sharding on string keys later.